fix: bugfixes for the encryption migration (#131)
* fix: copy the whole database when encrypting, not just all collections
* fix: only perform encryption migration if a database file already exists

+++ src/main/kotlin/de/darkatra/vrising/discord/persistence/DatabaseConfiguration.kt
@@ -2,7 +2,6 @@ package de.darkatra.vrising.discord.persistence
 
 import de.darkatra.vrising.discord.BotProperties
 import de.darkatra.vrising.discord.migration.SchemaEntityConverter
-import de.darkatra.vrising.discord.migration.getNitriteMap
 import de.darkatra.vrising.discord.migration.listAllCollectionNames
 import de.darkatra.vrising.discord.persistence.model.converter.ErrorEntityConverter
 import de.darkatra.vrising.discord.persistence.model.converter.PlayerActivityFeedEntityConverter
@@ -11,6 +10,8 @@ import de.darkatra.vrising.discord.persistence.model.converter.ServerEntityConve
 import de.darkatra.vrising.discord.persistence.model.converter.StatusMonitorEntityConverter
 import org.dizitart.no2.Nitrite
 import org.dizitart.no2.NitriteBuilder
+import org.dizitart.no2.collection.Document
+import org.dizitart.no2.collection.NitriteId
 import org.dizitart.no2.exceptions.NitriteIOException
 import org.dizitart.no2.mvstore.MVStoreModule
 import org.dizitart.no2.store.StoreModule
@@ -41,49 +42,52 @@ class DatabaseConfiguration(
         fun buildNitriteDatabase(databaseFile: Path, username: String? = null, password: String? = null): Nitrite {
 
             // version 2.12.0 introduced database encryption at rest. the following code attempts to perform the migration if necessary
-            val firstFewBytes = databaseFile.inputStream().readNBytes(ENCRYPTED_MARKER.length).toString(StandardCharsets.UTF_8)
-            if (firstFewBytes != ENCRYPTED_MARKER) {
+            if (databaseFile.exists()) {
 
-                // if the automated migration was aborted while writing the files to disc, restore the backup
-                val unencryptedDatabaseBackupFile = Path.of(System.getProperty("java.io.tmpdir")).resolve("v-rising-bot.db.unencrypted")
-                if (unencryptedDatabaseBackupFile.exists()) {
-                    logger.info("Found an unencrypted backup of the database at: ${unencryptedDatabaseBackupFile.absolutePathString()}")
-                    unencryptedDatabaseBackupFile.copyTo(databaseFile, overwrite = true)
-                    logger.info("Successfully restored the backup. Will re-attempt the migration.")
+                val firstFewBytes = databaseFile.inputStream().readNBytes(ENCRYPTED_MARKER.length).toString(StandardCharsets.UTF_8)
+                if (firstFewBytes != ENCRYPTED_MARKER) {
+
+                    // if the automated migration was aborted while writing the files to disc, restore the backup
+                    val unencryptedDatabaseBackupFile = Path.of(System.getProperty("java.io.tmpdir")).resolve("v-rising-bot.db.unencrypted")
+                    if (unencryptedDatabaseBackupFile.exists()) {
+                        logger.info("Found an unencrypted backup of the database at: ${unencryptedDatabaseBackupFile.absolutePathString()}")
+                        unencryptedDatabaseBackupFile.copyTo(databaseFile, overwrite = true)
+                        logger.info("Successfully restored the backup. Will re-attempt the migration.")
+                    }
+
+                    logger.info("Attempting to encrypt the bot database with the provided database password.")
+
+                    // retry opening the database without encryption if we encounter an error
+                    val unencryptedDatabase = try {
+                        getNitriteBuilder(getStoreModule(databaseFile, null)).openOrCreate(username, password)
+                    } catch (e: NitriteIOException) {
+                        throw IllegalStateException("Could not encrypt the database.", e)
+                    }
+
+                    unencryptedDatabaseBackupFile.deleteIfExists()
+
+                    // create an encrypted copy of the existing database
+                    val tempDatabaseFile = Files.createTempFile("v-rising-bot", ".db")
+
+                    val encryptedDatabase = getNitriteBuilder(getStoreModule(tempDatabaseFile, password)).openOrCreate(username, password)
+                    for (collectionName in unencryptedDatabase.listAllCollectionNames()) {
+
+                        val oldCollection = unencryptedDatabase.store.openMap<NitriteId, Any>(collectionName, NitriteId::class.java, Document::class.java)
+                        val newCollection = encryptedDatabase.store.openMap<NitriteId, Any>(collectionName, NitriteId::class.java, Document::class.java)
+
+                        oldCollection.entries().forEach { entry -> newCollection.put(entry.first, entry.second) }
+                    }
+                    unencryptedDatabase.close()
+                    encryptedDatabase.close()
+
+                    databaseFile.copyTo(unencryptedDatabaseBackupFile)
+                    tempDatabaseFile.copyTo(databaseFile, overwrite = true)
+
+                    unencryptedDatabaseBackupFile.deleteIfExists()
+                    tempDatabaseFile.deleteIfExists()
+
+                    logger.info("Successfully encrypted the database.")
                 }
-
-                logger.info("Attempting to encrypt the bot database with the provided database password.")
-
-                // retry opening the database without encryption if we encounter an error
-                val unencryptedDatabase = try {
-                    getNitriteBuilder(getStoreModule(databaseFile, null)).openOrCreate(username, password)
-                } catch (e: NitriteIOException) {
-                    throw IllegalStateException("Could not encrypt the database.", e)
-                }
-
-                unencryptedDatabaseBackupFile.deleteIfExists()
-
-                // create an encrypted copy of the existing database
-                val tempDatabaseFile = Files.createTempFile("v-rising-bot", ".db")
-
-                val encryptedDatabase = getNitriteBuilder(getStoreModule(tempDatabaseFile, password)).openOrCreate(username, password)
-                for (collectionName in unencryptedDatabase.listAllCollectionNames()) {
-
-                    val oldCollection = unencryptedDatabase.getNitriteMap(collectionName)
-                    val newCollection = encryptedDatabase.getNitriteMap(collectionName)
-
-                    oldCollection.values().forEach { document -> newCollection.put(document.id, document) }
-                }
-                unencryptedDatabase.close()
-                encryptedDatabase.close()
-
-                databaseFile.copyTo(unencryptedDatabaseBackupFile)
-                tempDatabaseFile.copyTo(databaseFile, overwrite = true)
-
-                unencryptedDatabaseBackupFile.deleteIfExists()
-                tempDatabaseFile.deleteIfExists()
-
-                logger.info("Successfully encrypted the database.")
             }
 
             return getNitriteBuilder(getStoreModule(databaseFile, password)).openOrCreate(username, password)
